refactor(cli): move parser into args.py, rename to main.py, add headings (#172)
