+++ src/lintquarto/__main__.py
@@ -1,6 +1,6 @@
 """Thin wrapper which just imports and calls the entry-point."""
 
-from .cli import main
+from .main import main
 
 if __name__ == "__main__":
     raise SystemExit(main())
