+++ src/lintquarto/main.py
@@ -6,7 +6,7 @@ import shutil
 import sys
 from pathlib import Path
 
-from .args import CustomArgumentParser
+from .args import CustomArgumentParser, build_parser
 from .config import load_config
 from .merge import merge_config
 from .processing import (
@@ -16,6 +16,10 @@ from .processing import (
     validate_no_commas,
 )
 from .registry import Formatters, Linters
+
+# ============================================================================
+# Main function: entry point for the lintquarto CLI.
+# ============================================================================
 
 
 def main() -> None:
@@ -65,105 +69,9 @@ def main() -> None:
     sys.exit(exit_code)
 
 
-def build_parser() -> CustomArgumentParser:
-    """
-    Create and configure the CLI argument parser.
-
-    Returns
-    -------
-    parser : CustomArgumentParser
-        CLI argument parser.
-    """
-    linters = list(Linters().supported.keys())
-    formatters = list(Formatters().supported.keys())
-
-    # Set up custom argumentparser with help statements
-    parser = CustomArgumentParser(
-        description="Lint Python code in Quarto (.qmd) files.",
-        epilog=(
-            "Configuration can also be provided in pyproject.toml under "
-            "[tool.lintquarto]. CLI arguments override configuration file."
-        ),
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
-
-    # Subcommands
-    subparsers = parser.add_subparsers(
-        title="commands",
-        dest="command",
-        required=False,
-    )
-    subparsers.add_parser(
-        "list",
-        help="List supported linters and whether they are available.",
-    )
-
-    # Default commands
-    parser.add_argument(
-        "-l",
-        "--linters",
-        nargs="+",
-        required=False,
-        choices=linters,
-        metavar="LINTER",
-        help=f"Linters to run. Valid options: {linters}",
-    )
-    parser.add_argument(
-        "-f",
-        "--formatters",
-        nargs="+",
-        required=False,
-        choices=formatters,
-        metavar="FORMATTER",
-        help=f"Formatter to run. Valid options: {formatters}.",
-    )
-    parser.add_argument(
-        "-p",
-        "--paths",
-        nargs="+",
-        required=False,
-        help="Quarto files and/or directories to lint.",
-    )
-    parser.add_argument(
-        "-e",
-        "--exclude",
-        nargs="*",
-        default=[],
-        metavar="[exclude_paths]",
-        help="Files and/or directories to exclude from linting.",
-    )
-    parser.add_argument(
-        "-n",
-        "--lint-non-exec",
-        action="store_true",
-        help="Also lint non-executable Python code chunks",
-    )
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        action="store_true",
-        help="Verbose output.",
-    )
-    parser.add_argument(
-        "-k",
-        "--keep-temp",
-        action="store_true",
-        help="Keep temporary .py files after linting.",
-    )
-    parser.add_argument(
-        "-c",
-        "--custom-commands",
-        action="append",
-        default=[],
-        metavar="COMMAND",
-        help=(
-            "Custom command to run against the generated .py file. "
-            "Repeat for multiple commands. "
-            'Example: --custom-commands "mytool"'
-        ),
-    )
-
-    return parser
+# ============================================================================
+# Helpers which validate args and extract and validate custom commands
+# ============================================================================
 
 
 def validate_args(
@@ -287,6 +195,11 @@ def parse_custom_commands(
     return custom_commands
 
 
+# ============================================================================
+# List tools
+# ============================================================================
+
+
 def list_tools() -> None:
     """Print all supported tools and whether they're available."""
     print()  # Blank line
@@ -303,6 +216,11 @@ def list_tools() -> None:
             flag = "✓" if status["available"] else "✗"
             print(f"  {flag} {status['name']:16s} - {status['message']}")
         print()  # Blank line
+
+
+# ============================================================================
+# Run formatter, linter or custom commands
+# ============================================================================
 
 
 def run_formatter(
